+++ maven.tienda/src/main/java/com/castelaofpe/maven/tienda/models/Categoria.java
@@ -14,17 +14,17 @@ public class Categoria {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id = 0L;
+	private Integer id = 0;
 	
 	@Column(name="name")
 	private String nombre = ""; 
 	
 	
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	} 
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
